chore(repo): add package as acceptable scope to commitlint for gk-lockfile
greenkeeper-lockfile currently has the hardcoded commit message of "chore(package)" which is failing
because of our commitlint rules. This adds "package" as an acceptible scope so greenkeeper-lockfile
will succeed. Ideally greenkeeper-lockfile should be updated to support custom commit messages (this
is logged as a bug) but this is a band-aid in the meantime.

brandingbrand-source-id: 8ff1e25bffab81f0043d8d3020f73c57f63adcaf