feat(pirateship): implemented promo products on landing screen
This adds a promo/featured products carousel on the Shop landing screen in PirateShip. The category ID is specified in the "promoProducts" property in the data source environment configuration.

While implementing, the Shop screen was migrated from using providers to using Redux directly. A new promoProducts reducer and data loader was added.