fix(flagship): updates/fixes to init script
There appears to be a bug in how node-suspect, the library we're using to interact with the command line during the init process, handles output buffers. The symptom is hanging when running init while using Node 8. For whatever reason, forcing react-native link into --verbose mode resolves this issue.

This also removes an extraneous sendEof() command that's not necessary when CodePush is not enabled for a project.