feat(flagship): add module install script for react-native-firebase
This adds a script to perform the necessary actions to install base Firebase and Firebase Analytics
on iOS and Android. As with the other module scripts, it'll be run only when a particular dependency
exists on a project, in this case react-native-firebase.

Requires the following to be added to the end project's config:

{
  firebase: {
    ios: {
      googleServicesPlistFile: '<path to GoogleService-Info.plist>'
    },
    android: {
      googleServicesJsonFile: '<path to google-services.json>'
    }
}

brandingbrand-source-id: e1bbae6f9c70414e39c487031c2fa816f1343da1