feat(flagship): improve add-keys-ios output
Added logging before an import occurs to clearly see what
certficate/identity is being imported as well as the output from the
`exec` calls to see the status.

Creating a keychain if one exists with the same name will actually crash
the script, so wrapped that in a `try-catch` block. This shouldn't be a
problem in CI environments, but allowing to use an already existing
keychain will be useful to those running the script locally.

Also wrapped the `copyFileSync` command in a `try-catch` block so as to
output a message similar to the security commands.