feat(fscomponents): update how fieldsoptions are extended in addressform
BREAKING CHANGE: This changes how fieldsOptions in AddressForm is extended via props from assignment
to merge. With the previous configuration, users would need to override the entire configuration for
each field in order to simply add or modify one option. This change may lead to side effects in apps
that already pass custom fieldsOptions to AddressForm.

brandingbrand-source-id: e656e1305aaee28a187dfffb1308fd9af797dd02