fix(pirateship): Allow pirateship to work outside flagship
With these changes, a user should be able to copy the pirateship package from flagship and immediately be able to make it function with minimal changes. I wasn't able to get the rn-cli.config.js file working in both places at this time, but removing the references to parent directories ("../..") and the blacklist should be the only changes needed after copying.