fix(fsapp): devMenu path override and web screen wrapper update
Allows setting devMenuPath in the config to override the path that the dev menu is displayed on web. Also updates the web screen wrapper to more-closely match the native version's functionality and will suppress the link to the dev menu with the same logic used to determine if the screen should be added to the list.