feat(fscomponents): custom styling for swatch show more/less
Allows setting maxSwatches to 0 or negative to continue showing all swatches even when show less is
selected. This also allows specific styling for the container when less or more is selected. This
does slightly alter the behavior because the style is applied to a view wrapping just the swatches
and not the more/less selection.

brandingbrand-source-id: ed15cb1496534a4a39908d81010c509af378eaef