fix(fsapp): updates to server-side handling of fsapp
The old version needed to be used in a way where it created a new instance of FSApp on every
request. This way you can create one instance of FSApp and call getApp with the config needed for
that request and it will just make a new redux store for each request.

brandingbrand-source-id: 10c801f17865644a27dd5ef79f9b2e084d03f272