refactor(fscomponents): ccard validation to remove moment
This refactors the validation in the CreditCardForm component to eliminate the
need for the moment library. Because we were using moment to simply grab the
month and date off a MM/YY formatted string it was easy to change. This helps
us achieve our goal of removing dependencies to moment because it's a huge
library.

brandingbrand-source-id: 1ec123a3567b8e17e200b1dc65bd3d0bfb00766d