chore: disable shallow clone for publish stage of ci
Lerna publish is failing on Travis with invalid tag names, and my suspicion based on some Github comments is that Lerna is unable to properly read the history because Travis defaults to shallow clones of repos.

This turns of shallow cloning for the deploy stage to attempt to resolve this issue. https://docs.travis-ci.com/user/customizing-the-build#git-clone-depth