feat(fsapp): Add functions to convert path props to component props
Also fixes an issue where path-to-regexp will encode part of a path before injecting it, then complain that it doesn't match.

Usage:
Add static functions matchConvert and urlConvert to your screen.
matchConvert will take in the props generated by path-to-regexp from your path and return the props that should be passed to your screen.
urlConvert does the opposite.
This allows you to parse out more-complicated data structures into and out of the path.