chore: remove xcode build from push ci step
This removes the Xcode build phase of CI when code is pushed to master. This has been erroring due to issues with the PirateShip provisioning profile. After consideration, I decided that this was unnecessary as we already perform an iOS compile on pull requests so this shouldn't be needed as long as we follow proper protocol in requiring CI to pass before merging pull requests.