feat(fsapp): track whether the drawer is visible
"isOpen" only tracks whether the drawer is open or closed, regardless of the animation transition.
This adds a "drawerVisible" property for the wrapped component, which will allow the component to
know whether it is displaying or not. This allows the component to, for example, hide itself when it
isn't visible, to avoid being tabbed to.

brandingbrand-source-id: b3121e8fb3767bf230d21c9f812bb884bd5bfe99