feat(fsproductindex): Add functionality for loading previous pages
Currently you can load more, but if you start on a later page, you can not load previous pages. This adds that as an additional option and also adds a minPage value to Pageables, so that it can track the range of pages loaded with load more or load previous. Functionality should be unchanged for pages without load previous turned on or that don't look at the minPage value.