chore(repo): pin react-native-permissions version
react-native-permissions released a breaking change with 1.2.0 (way to semver guys) that removes
support for React Native's built in Async Storage in favor of having a peer dependency for the newly
released react native community async storage. Because we do not currently have a dependency on the
new async storage package this causes an error.

This PR pins react-native-permissions to the last good version until we have time to properly add
support for the new async storage package.

brandingbrand-source-id: abde20a830821f0a5c477e06c0753fb4ff05d005