Merge develop into rnn2 (#952)
* fix(flagship): add targeted device to both Debug and Release config

* feat: Write Reviews Baazarvoice implementation

* refactor: move denormalize function to separate file

Keeping BazaarvoiceDataSource.ts clean

* feat: Write review submission typings

* fix: update react-native-htmlview to version 0.14.0

* feat(fsengagement): add new carousel option and full screen background

* chore(release): publish v4.1.0

* chore: update react native cli dependency to 1.11.2

This should resolve an issue we're seeing in CI as described here: https://github.com/react-native-community/cli/issues/586

* feat(fscomponents): add some option props to zoomCarousel

**Description**
added and implemented the following optional props:
* hideZoomButton
    * A bool to show/hide zoom functionality
* containerStyle
    * StyleProp to ensure that all wrapping elements are styled as desired
* renderImageWeb
    * A function that returns a react node

* feat(fscomponents): add nextArrowOnBlur prop to pass through

* feat(fscomponents): extend changes to non web-specific zoomcarousel
change containerStyle to fillContainer: bool

* chore: pin react-native-permissions version

react-native-permissions released a breaking change with 1.2.0 (way to semver guys) that removes support for React Native's built in Async Storage in favor of having a peer dependency for the newly released react native community async storage. Because we do not currently have a dependency on the new async storage package this causes an error.

This PR pins react-native-permissions to the last good version until we have time to properly add support for the new async storage package.

* fix(fsengagement): update react-native-video

there are compile issues of 3.X react-native-video
when compiling with react-native 0.59
version 4.x has better support with this react version

* fix(fsapp): Put dev menu as first screen

When there is a collision between the path matching for other screens and the dev menu, this will make the dev menu take priority.

* chore: remove add-keys scripts in favor of more generic versions

BREAKING CHANGE: This removes the add-keys-ios and add-keys-android scripts as they required an internal BB workflow to work. The new scripts, add-keys-ios-internal and add-keys-android-internal, can be configured with the environment files with passwords being passed in as environment variables.

For usage instructions, see the new wiki article: https://github.com/brandingbrand/flagship/wiki/Signing-Your-Apps

* chore: pin react-native-sensitive-info to 5.4.x

* feat(fscomponents): support more styling options for MultiCarousel

* feat(fscomponents): add hideZoomButton props to ZoomCarousel

* fix: Add 'Marquis' feature to Button with title dynamically reflecting CTA state

* chore(fscomponents): add missing typedef to modalhalfscreen component

* fix(flagship): update string replace to work in new fastfile

* fix: get default env from initial script

* build: upgrade deprecated webpack dev server

* fix(flagship): fix fastlane build lane

Added missing xcargs flags that tells xcode to make an unprovisioned build when "fastlane build" is executed.

* chore: tell greenkeeper to ignore more dependencies
react, metro-react-babel-preset, and babel should only be updated as part of coordinated React Native upgrades.

* fix: Add itemCount to cart interface in fscommerce (#877)

* fix: Modify the podspec version to ve 2.6.2 (#878)

* feat(fsshopify): account data source implementation

* build: upgrade deprecated webpack dev server

* fix: remove unnecessary checks

* fix(pirateship): Remove duplicate "name" in config (#867)

* chore: Update react-native-masked-text

* fix: exclude kotlin intermediate build files 

3rd party libraries begin using Kotlin and intermediate`*.kotlin_module` files are causing build errors. Added this rule to exclude those files

* chore: update react-dom library

* chore: Update react-native-touch-id dependency

* fix(flagship): Fix renaming edge case for android (#898)

The only file/directory with "flagship" in the name was the package
directory, which is already being renamed when fixing the package
directory structure.

Resolves https://github.com/brandingbrand/flagship/issues/895

* fix(fsweb): Use yarn.lock for fsweb (#906)

Allows us to lock versions of subdependencies when running fsweb.

* chore(release): publish v7.4.2

* fix(fsweb): Lock babel runtime in package.json (#905)

package-lock.json doesn't get published by npm, so this removes it and locks the version in package.json. This still leaves us open to other package updates breaking web, so I will be following up to see if yarn.lock doesn't have this issue.

* chore(release): publish v7.4.3

* chore: fix merge confict

* chore: fix merge confict

* feat(fscomponents): add custom modal content logic and test view

* feat(fscomponents): add condition to render custom test modal content

* Update README.md

* chore(release): publish v7.5.0

* chore(release): publish v7.5.1

* chore(release): publish v7.5.2

* chore(release): publish v7.5.3

* chore(release): publish v7.5.4

* chore(pirateship): update react-native-svg library (#889)

* chore: tell greenkeeper to ignore react and react native types

These types should be updated in sync with Flagship updates to the React and React Native libraries.

* feat(fsshopify):  add safe undefined handler

* feat(fsshopify):  add safe undefined handler

* feat(flagship): Add option to only include the default env (#925)

* feat(flagship): Add option to only include the default env

* chore(release): publish v7.6.0

* refactor: CreditCardNumber to avoid getDerivedStateFromProps (#926)

* chore: Update generating info plist (#924)

* feat(flagship): deprecate appcenter.apiconfig configuration option

This deprecates the appcenter.apiConfig configuration option in favor of using the APPCENTER_API_TOKEN environment variable. This makes the key optional in the typedef and updates the Fastfiles for iOS and Android such that not including this option will not break the builds.

* chore(release): publish v7.7.0

* chore: update dev dependencies to alleviate security issues (#932)

This updates commitlint, jest, lerna, and typedoc to their latest versions. Because some of the concerns were not yet addressed by their parent packages, I also had to add custom resolutions for handlebars and https-proxy-agent.

* chore(release): publish v7.7.1

* fix(fsbazaarvoice): fix review id normalizer (#936)

review.id is normalized from `ProductId` which is not unique, on the contrary, the values are usually the same because the reviews are usually from the same product. `Id` provides unique values so using that instead

* Update typedoc to the latest version 🚀 (#922)

* chore: update typedoc to version 0.15.1

* Update yargs in group flagship to the latest version 🚀 (#929)

* fix: update yargs to version 15.0.1

* Update husky to the latest version 🚀 (#930)

chore: update husky to version 3.1.0

* chore: add error message if user runs tsc on top level (#944)