fix(flagship): Updated modules and fixed pirateship Android
Jetifier doesn't run on hoisted modules, so this makes @brandingbrand/react-native-payments get updated. Also updated path-to-regexp to v6, react-native-svg to v10, react-native-webview to v8, react-native-video to 5, and removed the unused uglifyjs-webpack-plugin. Also updated all packages to the latest in-range version, excluding the @babel modules and modules that have it as a higher-versioned dependency, such as the @storybook modules.