feat(flagship): add ability to provide custom ios bundle version string
The CFBundleVersion property for iOS apps can now be customized by providing the ios.bundleVersion property in the selected env file. If one is not provided, the legacy behavior of using the package.json version number will be maintained. (Note that the legacy behavior changes the version number from xx.yy.zz to XXYYYZZZ, e.g. 3.4.2 to 03004002. If you use this new override, it will use whatever value you put in the env.)