fix(fscomponents): Updated Accordion component
There was an issue where the component would not update its size when the contents changed size. The most obvious case where this would fail is if you had two accordions nested with each other, opening/closing the inside one would not update the height of the outer one until you closed and opened it again. Also added the ability to put content inside of the accordion as a child instead of having to put it in the "content" prop and set title as a string instead of forcing a JSX.Element. This makes the component serializable as long as you avoid setting any styles. I also removed the titleHeight prop, because it didn't do anything, updated the comments to reflect the current state of the component, and updated the story and sample.