fix(fsweb): upgrade react-native-web to 0.12.2
React-native-web 0.12.2 is designed to be compatible with react-native >= 0.60

It did remove a few of the same things that got moved out of react-native, like the date picker,
webview, and async storage, so this aliases web versions of those. The react-native-community
version of async storage doesn't support web, so this just copies the version that react-native-web
used to have.

brandingbrand-source-id: 0a02e77e2c834a7192b3cf192b76bd379d9f7025