feat(flagship): remove hockeyapp and update appcenter for fastlane
This also removes the :beta lanes from the Fastfiles, as these attempted to upload builds to
Hockeyapp which has been shut down. The available lanes are :appcenter and :build for iOS and
:appcenter, :appcenter_bundle, and :build for Android.

BREAKING CHANGE: This removes the apiToken env config that could be used to configure App Center
uploads. This was previously deprecated in favor of passing in an environment variable instead:
APPCENTER_API_TOKEN.

brandingbrand-source-id: 21d69622c7538a68336551fda1d2794f4353ca11