fix(fsapp): close dev menu before restart
There is a bug in react-native-navigation where it can hang when it tries to close any open modals
when restarting.  This change closes the dev menu modal before trying to restart to avoid that
issue. It could still occur if another modal is open, though, so this is an incomplete fix.

TEST
------------------
- run on iOS
- open the dev menu and either change the environment or click "Reload JS"
- should restart the application

brandingbrand-source-id: d9fa5c65afc96b4d34325de0d0eb8ce8ad658a8a