fix(fscomponents): Remove redundant code in SearchModal
Almost all the code in SearchModal is identical to the code in SearchScreen, just with the Modal wrapper added and a few less options. This makes the SearchModal wrap a SearchScreen in a Modal, which allows those same options to work with SearchModal and removes repeated code. The result should be no changes to how the SearchModal works, except for the addition of some options in SearchScreen that weren't there before.

Also fixed some typing in SearchScreen and PirateShip's Search screen. brandPartSuggestions was referring to some old code that isn't part of fscommerce anymore, so that was removed as it is no longer part of the applicible types.