fix: resolve android build tools in subprojects
This adds a block to build.gradle to set the Android build tools version for subprojects (from external dependencies). This addresses a build error in release mode in which libraries are attempting to reference Material UI. This appears to be the community-supported resolution to this problem: https://stackoverflow.com/a/53371395

This also runs Jetify whenever init runs because this is a required step for builds to work.