feat: make permissions a soft dependency in fsengage
fsengage is a dependency on a number of fs packages, and fsengage had a hard dependency on react native permissions due to its geolocation handler for the core cms provider. react-native-permissions requires native setup and exposes itself to end users, which is not ideal given that most projects that use fsengage don't need geolocation functionality. This PR makes permissions a peer dependency and updates the geolocation handler to return false if react native permissions isn't available in the project.