feat(fsapp): Add additional options to SSR
Can now set options to call next() on certain screens, so middleware after the SSR can run. Also allows turning off rendering the app for the root path and overriding the html that is used as the base, in case the file is in a different location or you want to pass in some other string.