fix(fscomponents): make format hour test work in all time zones
The test as is fails in other time zones IE Mountain Time. This is because the `hours` specifiy times like 9PM but do not specify a time zone when when using new Date() it converts to the time zone of the user and thus falls back.

This changes moves the absolute time zones from the test to fix this behavior.

Signed-off-by: wSedlacek <wsedlacekc@gmail.com>