test(fscomponents): format hours test fixes
The test as is fails in other time zones IE Mountain Time. This is because the `hours` specifiy
times like 9PM but do not specify a time zone when when using new Date() it converts to the time
zone of the user and thus falls back.

This fixes two minor typos in the format hours testing

brandingbrand-source-id: 677b3af0e9f0b2f96f5fbee2ba9e3b39b96b341f