feat(fsapp): notFoundRedirect
When failing to find a matching path, this adds a fallback catchall route that will either redirect
to a given screen with setStackRoot or call popToRoot. Generally you will only use setStackRoot with
single screen apps, because it will make any tabs redirect to that same screen, while popToRoot will
go to the initial screen for a given tab.

brandingbrand-source-id: c19ba142bfb9955c9d31c895cfbe7979a10a019d