fix(fscomponents): fix web carousel only showing one image
This fixes an odd regression in which the Carousel component for web is only displaying one image.
The cause seems to be that using <View> causes react-native-web to create a <div> with a class
attached; this class seems to prevent the react-id-swiper library from binding to the slide. This
replaces the <View> with <div> which seems to resolve the problem.

Test:
1. Run `yarn dev:storybook`
2. Navigate to the Carousel story
3. Verify that the Carousel story displays multiple slides

brandingbrand-source-id: f86d1b49386bcf5dabf7cff0fc68a4af8e95b7fa