chore(repo): upgrade immer dependency for security
This upgrades Immer to resolve a security alert. Immer is a dependency that's required for React
Storybook through a couple different channels. This PR upgrades @storybook/react to the latest
version (not a breaking change) which removes the direct dependency to Immer. It also manually
resolves Immer to the latest patched version, because react-dev-utils (another dependency of
Storybook) has not updated their Immer dependency yet.

brandingbrand-source-id: 609c1e3af801ddcacc2eadc5ec7863614aa2c920