fix(fsapp): core fixes and refacoring after debuging
This fixes many issues that were found during the debugging process of `linked-app` including
exports, missing variables (ie path), dependncy issues, etc.

This also includes an initial albeit broken attempt at Modals for cross platforms. They are disabled
in the public interface as web is not currently in a working state but that will be fixed and
enabled in the public interface in a future update.

Signed-off-by: wSedlacek <wsedlacekc@gmail.com>

brandingbrand-source-id: aeb31bd3a6ac8c673c90084478ede296b02e61fd