chore(repo): fix travis builds
The build storybook step of CI was failing because fscomponents had a dependency on fsapp, and fsapp
has a dependency on project_env_index.js being set. project_env_index.js is created as part of the
init process. This adds init to the test stage to address the error.

brandingbrand-source-id: a7186a80b036bc9737d16aefea40fd356df7b95d