chore: add convenience scripts for publishing alpha versions
This adds a bash script to aide in publishing alpha versions. If no version is specified, it will parse the lerna.json file to determine the next alpha version. If a version is specified, it will use said version. In either case it will invoke Lerna with the default flags we want for publishing alpha versions (dist-tag, allow-branch, and exact).