feat(flagship): add navigation patch to support ios 15
iOS15 introduced a new appearance setting, scrollEdgeAppearance, that dictates the style of the tab bar when scroll content is aligned with the tab bar. React Native Navigation v6 does not handle this property, so the tab bar gets a black background when the user scrolls to the bottom of the screen.

This PR adds a patch script that patches react-native-navigation to apply the same tab bar styling to the scrollEdgeAppearance settings as well to allow for a consistent tab bar style regardless of scroll state.

This problem was fixed in react-native-navigation v7 but this allows us to continue supporting Flagship 10 and 11 which depend on v6.