fix(fscomponents): update web carousel dependencies
This updates the Swiper dependency to fix a security issue. The fixed version is a breaking change
from the previous (v5 to v6) which in turn broke our custom implementation which uses
react-id-swiper in conjunction with Swiper. (The pagination buttons were no longer displaying.)

I refactored the web carousel code to use the "custom build" solution described in the
react-id-swiper readme and it solved the problem:
https://www.npmjs.com/package/react-id-swiper#custom-build-swiper

This also adds a new screen to the dev menu to test carousels.

There are two ways to test this:

- Run `yarn dev:storybook` and play with the Carousel stories
- Run PirateShip web (`yarn ship:compile-web`), go to the develop menu, and then go to the Carousel
  sample screen

brandingbrand-source-id: 023cbf8c074d9ccb5a238e1d248f591d23fe79d9