feat(fscomponents): remove react-native-text-input-mask and dependent components
react-native-text-input-mask is a no-longer-maintained library that now has dependencies with
security vulnerabilities.

This PR also removes two components that depended on this library, CreditCardNumber and
CreditCardForm. These are legacy components with missing functionality that don't have external use.

It also refactors the Stepper components to manually parse the input so that we can remove the mask
as a dependency.

brandingbrand-source-id: 087436aa064bfada1865eb7f3528a4ccb6e25132