feat(fsmockdatasources): replace faker with own generated values
This removes Faker, a library used to generate fake data, from the repo as the maintainer has "rage quit" the project.

This implements a few random value generators including a simple lipsum sentence and paragraph generator.

Faker was only used to create fake reviews in the mock data source package.

brandingbrand-source-id: 365a300ade2518d0bb6fde5c39790ba841326495