feat(flagship-nx): add new generated add-keys-ios script
Part of the CI process for iOS in Flagship was the add-keys-ios script, which adds the required provisioning profile, cert, and key to the keychain. Previously this was handled using the environment file at runtime because these files were stored in the runtime config.

Now these configs are stored in the build configurations in project.json. As such, I created a generator in flagship-nx that will create the add-keys-ios script using the build configs. It's saved in the app's ios folder as this is not saved in version control.

brandingbrand-source-id: f6fada4265ee424cc3336d43396b0d829d25c2db