fix(fastlane): ENG-5249 android app center release mode
Android build number for is important for not only the Google
Play Store but also if you are installing subsequent versions
of a mobile application. The build number must be incremental
compared to the previous version installed.

Adding the conditional to release mode to always increment the
build nubmer based on the previously released version.