feat(adr): android 14 support
It as advised against updating agp/kotlin version outside of React Native
updates: https://github.com/facebook/react-native/issues/36974#issuecomment-1516829016.
Update to SDK 34 but supress error message for now as builds will continue
to work. It is advised to upgrade to Flagship Code™ v13.

Attempting to upgrade agp/kotlin versions result in a kotlin version mismatch to
compiled rnagp module - so upgrading is not possible:

The binary version of its metadata is 1.8.0, expected version is 1.6.0.

WARNING:We recommend using a newer Android Gradle plugin to use compileSdk = 34

This Android Gradle plugin (7.4.2) was tested up to compileSdk = 33

This warning can be suppressed by adding
    android.suppressUnsupportedCompileSdk=34
to this projects gradle.properties

The build will continue, but you are strongly encouraged to update your project to
use a newer Android Gradle Plugin that has been tested with compileSdk = 34